Uses Enumerator to create path
The path is passed to `visit` within each loop so the observer can see
the current node and the current path. However, path is expensive to
compute. Instead we pass `path` as an Enumerator. We create this
enumerator once as a closure capturing the current `node`.

This makes it efficient to pass `path` and only incur the cost of
generating the entire path when requested.

However, the receiver needs to recognize that the `path` is an
enumerator and must use `each` or other enumerable methods.
Alternatively, they can call `to_a` to create a realized array.